Move helper methods from `suppliers` to `shared_helpers`
These will be used by other views too, so better to have them in `shared`.
